Move plugin loading to root layout. Delay plugin loading message by 1 second.
